Refactor script.dart
